Code formatting only.
Ran sublime text's JSFormat extension to normalize whitespace,
brace style, ...   I'm going to put a lot of the functions/vars
in a module soon. So I thought that having another commit that
hits a lot of lines would not be too much of an inconvenience
for folks. I hope that is correct...