Explicitly forcing order of taxon-search results.
Since the order of results is not guaranteed, we check first for results matching the exact name, then any wildcard results.

We also now allow for more results (up to 100) in the dropdown menu, and I'm taking more care to remove redundant results (esp. an exact matching name, since this is liable to appear in both the exact and wildcard results).