Modified web2py auth to use OAuth2 from GitHub
This makes GitHub user credentials the only way to login to the curation
tool. (The Janrain auth used in the main 'opentree' app is unchanged,
since commenting on the tree shouldn't require a GitHub account.)

This needs more work to create/update the user's record in db.auth_user,
but it's the foundation for using the GitHub APIs to correctly act in
one's own name in the OTOL repo.