Adding (mostly) inactive curation UI, loads/saves
Most of the study editor is an HTML mockup. It's just functional enough for our next integration milestone, in that is loads a specified study's Nexson from the OTOL API, allows editing of a few metadata fields, and saves those changes back to storage.

Similarly, the "dashboard" page for a logged-in curator is just a sketch, but it has one live link ("Example study one") that opens the editor to the study with id=10.