First run at showing D3 tree using test .json file
This is based on earlier bare-bones tests. Note that the JSON in this case is in D3's expected format (nested node objects, each with a "children" array of more). To conserve memory, I'm hoping to teach D3 to understand a Nexson tree structure rather than re-building the tree inthis form.