First pass at removing KO mapping and observables
By removing observables within the 'nexml' element (Nexson data proper), we sidestep most of the performance penalties and browser crashes for large studies. The price we pay is to re-frame some logic (esp. for reading and writing Nexson values) and to set up a system of explicit "ticklers" to trigger recalculation and UI updates. This is in progress now, with working study metadata.