Adding fast lookup of Nexson elements by ID
The "fast lookup" feature is a major speed-up when analyzing tree structure, eg, when preparing for a tree display in d3. Note that these should be cleared and rebuilt explicitly if the underlying data changes.

I've also cleaned out some other obsolete code. Still working on tree-draw and -refresh performance for the very largest trees. Current focus is on optimizing the d3 phylogram script.