Sending curator to View page (vs. immediate Edit)
This is in response to concerns that it's too easy (or unnerving) for a
logged-in user to jump immediately into some other study and possibly
lose/mangle its data.

N.B. This does not change our simple rule that any logged-in user
(anyone with a GitHub account) can edit any study in the system.