Add toggling server-details box in app footers.
For now, this is a dump of all non-secure values in the
```private/config``` file for the current web2py app. For the curation
app, we also do an AJAX fetch (lazy) to fetch the list of all shards
(phylesyste repos) used in the current phylesystem-api instance, and a
study count for each.