If tree imports successfully, add supporting file.
This copies the input text (whether an uploaded file or pasted into a
form field) to a new file in the supporting-files area for this study,
so that it can be reviewed, downloaded, and deleted from within the
curation app. Fixes #418.

NOTE that this is in addition to the saving of input data to peyotl's
scratch folders. The latter is for diagnostic purposes.