Removed branch name from fetch.
Odd. Adding a branch name seemed to make the fetch
do nothing, but not exit with an error. I tested this
with a newly created branch in phylesystem-api, and the fetch
silently failed during deployment leading to a subsequent
fatal error when we tried create the branch. ssh-ing into
the dev machine showed that the origin/pseudo-branch had
not been fetched. Removing the branch name from the fetch
seemed to work.