Adapt viz to browser width, place details nicely.
This is a good-enough solution for now, but it doesn't take full
advantage of resizing after the page has loaded (viz scales up/down, but
doesn't show more information or separate markers nicely).

Later, when we have more significant dates to manage, we should consider
scrolling (or scaling) the viz timeline, and ideally taking more
advantage of dynamic resizing.