Simple caching (and clearing) of treeview comments.
This caches local comments indefinitely, but with explicit clearing
triggered by a webhook in the 'feedback' repo, defined here:
  https://github.com/OpenTreeOfLife/feedback/settings/hooks

This webhook is triggered by the 'Issues' and 'Issue comment' events; as
a result, adding/removing/updating/replying to an issue will clear all
cached comments. Once I have a better grip on the payload from GitHub
API, we should be able to selectively clear only the comments for the
affected location(s).