Clean up logic for tree-ranking UI.
This now handles up and down nudges correctly, and makes sense of
user-entered rank values even when out of range or otherwise bizarre.

The current behavior (after Move All, or a nudge) is to reset all Rank
values to simple series of integers: (1, 2, 3, ...)
If curators prefer to use arbitrary or forward-looking rank values, we
should try to preserve them as entered, and possibly to save them in the
stored collection as a complement to the "natural" order of the array
(which also reflects the latest ranking, but with no gaps in the array).