Add Edit button & explanation to curator profile
This appears if the current user is also the profiled curator. It
explains how the fields in a GitHub profile map to the displayed
information in the OpenTree curator profile, then tries to open their
GitHub user profile page in a new tab or window.

I've also removed the display code for the (deprecated) "Bio" on GitHub.