Adding a filter 'skip_comments' to just get markup
This gives us the UI and skips the (laggy) fetch of comments from
GitHub. Mainly useful for the synth-tree viewer page, which loads
"empty" ASAP and is quickly updated by Javscript.

Also fixed the (old) default filter in most page tempates to be the
expected string 'url' instead of a tuple. (This was "working
accidentally" on most pages, since the tuple was being ignored but the
default filter was 'url'.)