Manage depth (z-index) of stacked Bootstrap modals
This uses global even delegation on jQuery modals, so that we can
sensibly "stack" or chain these, for example:
 > main study editor page
 > modal tree-editor popup
 > modal event blocker (while loading collections list)

This ensures that each modal has its own event-blocking screen (visible
as an 80% black transparency); that these accumulate (darken) when we're
deep in the modal chain; and that each subsequent modal UI (and its
screen) appears on top of those prior. So we get the proper responses to
click events, etc. and enforce modality of the entire stack.