Change the tree-upload form validation.
This was using the value of the input[type=file] widget as part of form
validation; as it turns out, the upload widget clears this after each
file is added, so it's not a reliable test. Instead, we'll actually use
the filename preview, since this is maintained by the widget. (In each
test, we're looking for a non-empty field value *or* element text.)
Fixes #542.