Block Continue button if adding a duplicate DOI
This is more careful to withhold the Continue button while waiting for
fresh results. NOTE that there's one edge case that can be a little
confusing: if someone goes directly from editing the DOI field to the
Continue button, there's a momentary test before the button comes alive.

I don't see a way around this, since this gesture blurs the DOI field
(triggering a dupe-test) just as the user is trying to use the button.