Add conflict control/indicator to tree popup
This works properly, in the sense that it reflects the reference tree of
an incoming URL, and it allows changing (or removing) the conflict
display in the tree viewer. It doesn't currently work through the
browser History (should update the URL and let *that* trigger the viz,
as with the incoming URL).