OTU list should only show OTUs in use.
For now, we preserve any OTUs in the study even after trees are deleted.
This is to save curator time in case the same OTUs (original node
labels) are re-imported with a new tree. But the OTU list should hide
these unused OTUs even if the 'In all trees' setting is used.

Addresses #909, but doesn't answer the main question there.