Toggle "shared" evidence for new taxa (locks UI)
This provides the UI for sharing a single set of evidence fields (the
evidence type and one line of text) across all new taxa. This is
indicated by locking/disabling the evidence fields. N.B. the original
values for all OTUs are preserved and will re-appear if the curator
"un-shares" the chosen evidence.