Remove diacritic marks from query-string (safer).
This prevents encoding problems when diacritic (non-Latin) characters
are passed to the query string via History.js. It relies on the same
patterns that we use to include all diacritical variants in our
client-side searches and list filters, so the resulting "Latinized"
value will generate the same results as the original if the URL is
shared.