Put SVG stylesheet where Inkscape can find it.
It turns out [CSS support in Inkscape is
minimal](http://wiki.inkscape.org/wiki/index.php/CSS_Support), and it
only accepts a stylesheet as a top-level `svg:style` child of the main
`svg` element. The browsers seem OK with this change, so I'm guessing
Illustrator and others will be as well.