Update incoming study Nexson on-the-fly.
Update to new-style tree property `ot:readyForSynthesis`, versus list
stored on the main `nexml` element. Before discarding the old
`ot:candidateTreeForSynthesis` list, set any trees previously marked as
preferred to be included in the next synthesis.

N.B. This correctly captures trees marked as "preferred" in the old
system, but does not detect the prior inclusion of a tree in a synthesis
collection! This will require a batch operation or more client-side
smarts.