Rough fix for too-large labels
This is not ideal, since sometimes there's a visible glitch in the tree
layout as it resizes. Also, it appears we pass through the tree render a
few times, and (for the first tree) the bounding box is height=0,
width=0, so we need to suppress the resize. Still, this should fix the
unwanted cropping of trees in both screen and print.