Handle new-style synth-node IDs in conflict.
Avoid a zero-day situation by handling old- _and_ new-style node ids
returned from the conflict service for the synthetic tree. Addresses #1062.

Note that we'd ideally like to incorporate the latest synthesis release
in the resulting URL, e.g. 'opentree7.0'. This could be done by
including this string in the study-editor page, or by returning it as an
additional property from the conflict service.