Ignore tips when looking for ambiguous node labels
These labels are not really ambiguous! If found, a leaf node with a
`@label` attribute seems only to indicate a misspelled taxon name that
was entered elsewhere (i.e. imported from TreeBASE).

By reporting them here, we present the user with a confusing instruction
to disambiguate labels, but nothing they can see and respond to.