Allow more recent stevedore.
Not sure how to test if this change works. I see no
obvious problems with it. I don't even see where we
use this package within our webapps, so we can probably
get rid of it.

The version 0.8 that we were previously pinned to does
not install cleanly on any modern Python version that I've seen.