Initial commit of early/broken TNRS tool
This is adapted from prior work in the study curation and
Tree Illustrator apps. Still refactoring and streamlining code from
both.

N.B. This is a single-page app, with almost all logic in the browser. It
uses ZIP files to load/save sessions ("namesets") and has minimal
dependence on the web2py framework. I'm adding it to the existing
'curation' app in web2py primarily to take advantage of shared style,
UI, and existing curation-helper functions.