Fixes and improvements to collection-mgt UI
Update API calls to look up collections, to recognize the latest
searchable properties. This fixes some serious bugs in the current
curation tool (e.g. no way to add trees to a collection!).

Use the same API calls to add an 'Update trees' feature when editing a
collection. This will update the collection to store the latest tree
names/labels, and also mark any trees (or studies) that have been
removed from phylesystem, so that the curator can replace them if
needed. As usual, these operations are recorded and summarized in the
commit messages when saving the collection.