Fix an error in new bulk TNRS (no matches == null)
The previous version of TNRS returned an empty array for
`results.matches` if there were no matches found.

I've added a safety net to the JS to emulate the old response in the
curator app, but we should probably try to keep the old behavior if
there are any other API consumers out there.