Smarter logic for refreshing MODIFIED trees
When "refreshing" a tree collection, we check its listed trees against the matching trees and studies in phylesystem, specifically any properties that are also stored in a tree collection. If *any* of these properties have changed, update our listed tree and mark it as MODIFIED. Offer these changes to a human curator, as before.