Use '@' instead of '#' (else History breaks)
This is a slightly confusing hack. Suffice to say that the browser's
History (session history) ignores URL fragments signified by '#', so it
doesn't recognize when we choose a different target tree in the same
study. Encoding it as '%23' in the History state doesn't seem to matter,
as it's decoded when the address bar is updated.

By substituting '@' in our code and URLs, we should satisfy both
History and the conflict API, which expects to see '%23'