Allow for multiple "exact matches" in OTU mapping
This uncovers a faulty assumption in my original code, that any exact
match would be a singleton. Instead there are often a few candidates
whose names match exactly, but differ in other subtle ways.

These changes should properly auto-accept exact matches for OTUs in the
study curation webapp, for both of these controls:

- 'Auto-accept exact matches and synonyms' (a persistent setting for this study)
- 'Accept exact matches' (a button that triggers one-time auto-acceptance)

NB - This code is in the study editor. Once tested, it will be copied
into the bulk-TNRS code and its node module re-compiled.